+++ app/templates/layout.html
@@ -41,13 +41,14 @@
         {% endblock %}
     <script src="https://cdnjs.cloudflare.com/ajax/libs/flowbite/2.3.0/flowbite.min.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
-    <script defer src="https://cdn.jsdelivr.net/npm/alpinejs@3.13.5/dist/cdn.min.js"></script>
     <script type="module" src="{{ url_for('static', filename='js/toast.js') }}" defer></script>
-    <script src="{{ url_for('static', filename='js/tableInteractions.js') }}"></script>
+    <script type="module" src="{{ url_for('static', filename='js/tableInteractions.js') }}"></script>
     <script src="{{ url_for('static', filename='js/tableState.js') }}"></script>
     <!-- <script src="{{ url_for('static', filename='js/rowComponent.js') }}"></script> -->
     <script src="{{ url_for('static', filename='js/createRow.js') }}"></script>
-    <!-- <script src="{{ url_for('static', filename='js/textScramble.js') }}"></script> -->
+    <script type="module" src="{{ url_for('static', filename='js/sendRequest.js') }}"></script>
+    
+    <script defer src="https://cdn.jsdelivr.net/npm/alpinejs@3.13.5/dist/cdn.min.js"></script>
 
 </body>
